Fix Design Flaw in IWBSRegion
While coding today I tried to put a fluid container inside of a navbar. I discovered that it added extra attributes. I appreciate the attempt to try to handle the common usage of a container inside a navbar, but it doesn't work to force these attributes on a container, because it's sometimes necessary to have a fluid container without them.